% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodVirusData-class.R
\name{FoodVirusData}
\alias{FoodVirusData}
\title{An S4 class for representation of foodvirus-data}
\usage{
FoodVirusData(...)
}
\arguments{
\item{...}{A data frame or list to be converted into a \code{FoodVirusData}-object.}
}
\value{
A \code{FoodVirusData}-object if validation succeeds, an error
message otherwise.
}
\description{
The \code{FoodVirusData}-class stores verification and validation data,
which are used as input to the foodvirus Shiny application.
A \code{FoodVirusData}-object has a data-frame structure and contains
two numeric variables: dilution and obtained.

The class extends the \code{S4Vectors::DataFrame} class
(Pages et al., 2021), without any additional slots.
}
\examples{
## Constructor
data("oyster")
x <- as.data.frame(oyster)
FoodVirusData(x)

## Coercion to a traditional data-frame
as.data.frame(x)
as(x, "data.frame")
}
\references{
Pages, H., Lawrence, M., and Aboyoun, R. (2021). S4Vectors:
Foundation of vector-like and list-like containers in
Bioconductor. R package version 0.30.0.
}
