% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyster-dataset.R
\docType{data}
\name{oyster}
\alias{oyster}
\title{Example dataset for the foodvirus Shiny application}
\format{
An object of class \code{FoodVirusData} with 90 rows and 2 columns.
}
\usage{
data("oyster")
}
\description{
The `oyster`-dataset contains data from a validation study on oyster samples.
}
\keyword{datasets}
