% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{runEURLVerguide}
\alias{runEURLVerguide}
\title{Open the EURLVerguide application}
\usage{
runEURLVerguide()
}
\value{
Opens a Shiny application.
}
\description{
Open the EURLVerguide application
}
\examples{
## Only run this in interactive R sessions:
if (interactive()) {
    runEURLVerguide()
}
}
